Add game logic
Initially, I implemented the game logic using only switch case, using
both playerShape and computerPlay functions as arguments, but bugs
started to appear: When 'paper' was selected by the player, the game
would register that as 'rock'. This was solved using three ifs and
nesting switch cases inside them.
The initial logic is registered and commented for future reference.